# packages/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Bookmark Manager Platform API
  description: |
    Production-ready bookmark management platform with comprehensive organization, archiving, search, and collaboration capabilities.

    ## Features
    - Save and organize web bookmarks with rich metadata
    - Hierarchical collections with custom icons
    - Multi-tag support with advanced filtering
    - Full-text search within saved page content (Pro)
    - Permanent archival with page snapshots (Pro)
    - Highlights and annotations (Pro)
    - Collaboration and sharing (Pro)
    - Cross-device real-time synchronization

    ## Authentication
    The API supports two authentication methods:
    1. **JWT Bearer Tokens** - For direct user authentication
    2. **OAuth 2.0 with PKCE** - For third-party applications

    ## Rate Limiting
    - Free tier: 100 requests per minute per user
    - Pro tier: 500 requests per minute per user
    - Developer API: Rate limits based on application tier

    When rate limits are exceeded, the API returns HTTP 429 with `Retry-After` header.

  version: 1.0.0
  contact:
    name: API Support
    email: support@bookmarkmanager.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.bookmarkmanager.example.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login

    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Read access to bookmarks and collections
            write: Create and update bookmarks and collections
            delete: Delete bookmarks and collections
            admin: Full administrative access
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
            - requestId
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: INVALID_REQUEST
            message:
              type: string
              description: Human-readable error message
              example: Invalid request parameters
            details:
              type: object
              description: Additional error context
            timestamp:
              type: string
              format: date-time
              description: ISO 8601 timestamp
            requestId:
              type: string
              description: Unique request identifier for tracing
...
    OAuthClient:
      type: object
      required:
        - id
        - client_id
        - name
        - redirect_uris
        - is_public
        - created_at
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          description: OAuth client identifier
        client_secret:
          type: string
          description: Client secret (only returned on registration for confidential clients)
        name:
          type: string
          maxLength: 255
          description: Application name
        redirect_uris:
          type: array
          items:
            type: string
            format: uri
          description: Allowed redirect URIs
        is_public:
          type: boolean
          description: Whether this is a public client (PKCE required)
        created_at:
          type: string
          format: date-time
paths:
...
  /oauth/clients:
    post:
      tags:
        - OAuth
      summary: Register a new OAuth client application
      operationId: registerOAuthClient
      description: |
        Register a third-party application to access the API on behalf of users.
        Public clients (mobile apps, SPAs) should use PKCE flow.
        Confidential clients (server-side apps) receive a client secret.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - redirect_uris
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  description: Application name
                redirect_uris:
                  type: array
                  items:
                    type: string
                    format: uri
                  minItems: 1
                  description: Allowed redirect URIs for OAuth flow
                is_public:
                  type: boolean
                  default: true
                  description: Whether this is a public client (requires PKCE)
            example:
              name: My Bookmark App
              redirect_uris:
                - https://myapp.example.com/callback
              is_public: true
      responses:
        '201':
          description: OAuth client registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthClient'
              example:
                id: 550e8400-e29b-41d4-a716-446655440000
                client_id: a1b2c3d4e5f6g7h8
                name: My Bookmark App
                redirect_uris:
                  - https://myapp.example.com/callback
                is_public: true
                created_at: '2024-01-15T10:30:00Z'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /oauth/authorize:
    get:
      tags:
        - OAuth
      summary: OAuth 2.0 authorization endpoint
      operationId: authorizeOAuth
      description: |
        Initiates the OAuth 2.0 authorization flow.
        Users will be redirected to login/consent screen.
      parameters:
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
          description: PKCE code challenge
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256, plain]
        - name: scope
          in: query
          schema:
            type: string
          description: Space-separated list of scopes
        - name: state
          in: query
          schema:
            type: string
          description: Client state for CSRF protection
      responses:
        '200':
          description: Authorization page
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /oauth/token:
    post:
      tags:
        - OAuth
      summary: Exchange authorization code for access token
      operationId: exchangeToken
      description: |
        Exchange an authorization code for access and refresh tokens.
        Requires PKCE code verifier for public clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grant_type
                - code
                - client_id
                - redirect_uri
                - code_verifier
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code]
                code:
                  type: string
                  description: Authorization code from /oauth/authorize
                client_id:
                  type: string
                redirect_uri:
                  type: string
                  format: uri
                code_verifier:
                  type: string
                  description: PKCE code verifier
            example:
              grant_type: authorization_code
              code: abc123def456
              client_id: a1b2c3d4e5f6g7h8
              redirect_uri: https://myapp.example.com/callback
              code_verifier: dBjftJeZ4CVP-mB92K27uhbUJU1p1r_wW1gFWFOEjXk
      responses:
        '200':
          description: Token issued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  token_type:
                    type: string
                    example: Bearer
                  expires_in:
                    type: integer
                    description: Token lifetime in seconds
                    example: 900
                  scope:
                    type: string
                    description: Space-separated list of granted scopes
              example:
                access_token: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...
                refresh_token: def456ghi789jkl012mno345pqr678stu901
                token_type: Bearer
                expires_in: 900
                scope: read write
        '400':
          description: Invalid request or grant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...